refactor: group pi package lookup support
